facets: experiments facets fix
* Experiments matching now also support multiple experiments for a record.

* Experiments facets are now correctly displayed.

* Closes #792 .
