+++ inspirehep/base/searchext/mappings/hep.json
@@ -103,6 +103,10 @@
                             "type": "string",
                             "index": "not_analyzed",
                             "copy_to": ["experiment"]
+                        },
+                        "facet_experiment": {
+                            "type": "string",
+                            "index": "not_analyzed"
                         }
                     }
                 },
