+++ inspirehep/base/templates/search/facets/hep.html
@@ -126,12 +126,12 @@
     {% for experiment in aggs['experiment']['buckets'] %}
     <div class="checkbox checkbox-facets">
       <label>
-        {%- set is_checked = 'checked' if 'experiment' in filtered_facets
-            and experiment['key'] in filtered_facets['experiment']['inc'] -%}
+        {%- set is_checked = 'checked' if 'experiment_f' in filtered_facets
+            and experiment['key'] in filtered_facets['experiment_f']['inc'] -%}
         <input class="include-facet"
               type="checkbox"
               {{is_checked}}
-              name="experiment" value="{{ experiment['key'] }}">
+              name="experiment_f" value="{{ experiment['key'] }}">
         <span class="facet-title-truncate">{{ experiment['key'] }}</span>
         <span class="facet-label label label-default {{ 'label-primary' if is_checked}} pull-right">{{ experiment['doc_count'] }}</span>
       </label>
